Move logSchema to consts for client-side validation
